extract visualizer lib/package
