+++ libs/react-rx-store/src/lib/react-rx-store.tsx
@@ -27,15 +27,14 @@ export const useStore = <T extends {}>(context: Context<T>): T => {
   return value;
 };
 
-
 declare global {
   interface Window {
     __rxStoreSubjects: Subject<{
-      name: string
+      name: string;
     }>;
     __rxStoreEffects: Subject<{
-      name: string,
-      event: 'spawn'|'teardown'
+      name: string;
+      event: 'spawn' | 'teardown';
     }>;
   }
 }
@@ -46,7 +45,7 @@ declare global {
 // const before$ = (ref: { name: string }) =>
 //   of(null).pipe(
 //     tap(() => {
-      
+
 //     }),
 //     filter(() => false)
 //   );
@@ -58,10 +57,10 @@ const after = (ref: { name: string }) => {
   // TODO - add a devtools hook here
   debug(`rx-store:${ref.name}`)('teardown');
   // TODO this won't cleanup on unsubscribe!
-  window.__rxStoreEffects.next({...ref, event: 'teardown'});
+  window.__rxStoreEffects.next({ ...ref, event: 'teardown' });
 };
 
-const ids = {}
+const ids = {};
 
 /**
  * The spawnRootEffect runs the root effect which means the effectFn is called
@@ -105,32 +104,32 @@ export const spawnRootEffect = <T extends {}>(
     // Keeps the "context" intact, by appending to debugKey to create a path
     // each time an effect creates a child effect by running it's curried `spawnEffect()`
     const childSpawnEffect: SpawnEffect<T> = (effect, { name: childName }) => {
-      const curriedName = `${name}:${childName}`
+      const curriedName = `${name}:${childName}`;
       if (undefined === ids[curriedName]) {
         ids[curriedName] = 1;
       } else {
         ids[curriedName]++;
       }
       const id = ids[curriedName];
-    
-      ensureDevtools()
-      
+
+      ensureDevtools();
 
       debug(`rx-store:${curriedName}:${id}`)('spawn');
 
-      window.__rxStoreEffects.next({name: `${curriedName}:${id}`, event: 'spawn'});
+      window.__rxStoreEffects.next({
+        name: `${curriedName}:${id}`,
+        event: 'spawn',
+      });
 
       window.__rxStoreLinks.next({
-        from: {type:'effect', name:name},
-        to: {type:'effect', name:`${curriedName}:${id}`},
-      })
+        from: { type: 'effect', name: name },
+        to: { type: 'effect', name: `${curriedName}:${id}` },
+      });
 
       return spawnEffect(effect, {
         name: `${curriedName}:${id}`,
       });
     };
-
-
 
     // Run the effect function passing in the curried sources, sinks, and
     // spawnEffect function for the effectFn to run any of its children effectFn
@@ -138,12 +137,12 @@ export const spawnRootEffect = <T extends {}>(
 
     // Sandwich the effect between before and after streams, allowing devools
     // hooks to run when the effect is subscribed & torn down.
-    
-    return effect$.pipe(finalize(() => after({name})));
+
+    return effect$.pipe(finalize(() => after({ name })));
   };
 
-  ensureDevtools()
-  window.__rxStoreEffects.next({name: `root`, event: 'spawn'});
+  ensureDevtools();
+  window.__rxStoreEffects.next({ name: `root`, event: 'spawn' });
   return spawnEffect(rootEffectFn, { name: 'root' });
 };
 
@@ -162,6 +161,11 @@ export const store = <T extends {}>(
 ): { Manager: React.ComponentType<{}>; context: Context<T> } => {
   /** Each store gets a React context */
   const context = createContext<T>(value);
+
+  ensureDevtools();
+  Object.keys(value).forEach((name) => {
+    window.__rxStoreSubjects.next({ name });
+  });
 
   /**
    * This Manager must be mounted at most once, wrap your children
@@ -189,10 +193,7 @@ export const store = <T extends {}>(
       if (!rootEffect) {
         return null;
       }
-      const subscription = spawnRootEffect(
-        value,
-        rootEffect
-      ).subscribe();
+      const subscription = spawnRootEffect(value, rootEffect).subscribe();
       return () => subscription.unsubscribe();
     }, [value]);
